benchmark-performance-for-classical-ml-models/requirements.txt
black==22.10.0
click==8.1.3
cmdstanpy==1.0.8
coloredlogs==15.0.1
contourpy==1.0.6
convertdate==2.4.0
cycler==0.11.0
Cython==0.29.32
decorator==5.1.1
dill==0.3.6
ephem==4.1.3
flake8==5.0.4
flatbuffers==22.10.26
fonttools==4.38.0
hijri-converter==2.2.4
holidays==0.17
humanfriendly==10.0
hummingbird-ml==0.4.6
isort==5.10.1
joblib==1.2.0
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
lightgbm==3.3.3
LunarCalendar==0.0.9
matplotlib==3.6.2
mccabe==0.7.0
mpmath==1.2.1
mypy==0.991
mypy-extensions==0.4.3
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
onnx==1.12.0
onnxconverter-common==1.13.0
onnxruntime==1.13.1
packaging==21.3
pandas==1.5.1
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.5.4
plotly==5.11.0
prophet==1.1
protobuf==3.20.1
psutil==5.9.4
pycodestyle==2.9.1
pyflakes==2.5.0
PyMeeus==0.5.11
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
scikit-learn==1.1.1
scipy==1.9.3
setuptools-git==1.2
six==1.16.0
skl2onnx==1.13
sympy==1.11.1
tenacity==8.1.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.0
tqdm==4.64.1
treelite==3.0.0
treelite-runtime==3.0.0
types-decorator==5.1.8.1
typing-extensions==4.4.0
xgboost==1.6.2
